+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - We've added a logo now, and hidden many of the logs. These can be shown with the
   `--verbose` flag.
 
+### Fixed
+
+- Fixed a bug related to the parsing of the PDFs of some papers, due to the file not
+  ending in `.pdf`. This has been fixed now.
+
 ## [v0.1.2] - 2025-10-09
 
 ### Fixed